feat : 전체주문조회기능 추가

+++ backend/src/main/java/com/ll/cafeservice/entity/order/Order.java
@@ -41,7 +41,7 @@ public class Order extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private OrderStatus status;
 
-    private double totalPrice;     // 총 가격
+    private long totalPrice;     // 총 가격
 
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
     @JsonManagedReference
